+++ database/migrations/2021_03_05_200841_buat_pembelian_table.php
@@ -18,7 +18,7 @@ class BuatPembelianTable extends Migration
             $table->integer('id_supplier');
             $table->integer('total_item');
             $table->integer('total_harga');
-            $table->tinyInteger('diskon')->default(0);
+            $table->integer('diskon')->default(0);
             $table->integer('bayar')->default(0);
             $table->timestamps();
         });
