+++ database/migrations/2021_03_05_200858_buat_penjualan_detail_table.php
@@ -19,7 +19,7 @@ class BuatPenjualanDetailTable extends Migration
             $table->integer('id_produk');
             $table->integer('harga_jual');
             $table->integer('jumlah');
-            $table->tinyInteger('diskon')->default(0);
+            $table->integer('diskon')->default(0);
             $table->integer('subtotal');
             $table->timestamps();
         });
